+++ src/main/java/roomescape/controller/AdminViewController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 @RequestMapping("/admin")
-public class AdminController {
+public class AdminViewController {
 
     @GetMapping
     public String mainPage() {
